+++ Example/ViewController+FreeDrive.swift
@@ -43,7 +43,6 @@ extension ViewController {
         
         if let location = notification.userInfo?[PassiveLocationManager.NotificationUserInfoKey.locationKey] as? CLLocation {
             trackStyledFeature.lineString.coordinates.append(contentsOf: [location.coordinate])
-            navigationMapView.moveUserLocation(to: location)
         }
         
         if let rawLocation = notification.userInfo?[PassiveLocationManager.NotificationUserInfoKey.rawLocationKey] as? CLLocation {