+++ Sources/MapboxCoreNavigation/EHorizon/OpenLRIdentifier.swift
@@ -1,30 +1,30 @@
 import Foundation
 import MapboxNavigationNative
 
-/** Standard of OpenLR */
-public enum OpenLRStandard {
+/** Identifies a road object according to one of two OpenLR standards. */
+public enum OpenLRIdentifier {
 
     /**
      [TomTom OpenLR](http://www.openlr.org/).
 
      Supported references: line location, point along line, polygon.
      */
-    case tomTom
+    case tomTom(reference: RoadObjectIdentifier)
 
     /**
      TPEG OpenLR.
 
      Only line locations are supported.
      */
-    case tpeg
+    case tpeg(reference: RoadObjectIdentifier)
 }
 
 extension MapboxNavigationNative.OpenLRStandard {
-    init(_ standard: OpenLRStandard) {
-        switch standard {
-        case .tomTom:
+    init(identifier: OpenLRIdentifier) {
+        switch identifier {
+        case .tomTom(_):
             self = .tomTom
-        case .tpeg:
+        case .tpeg(_):
             self = .TPEG
         }
     }