+++ Sources/MapboxCoreNavigation/NavigationEventsManager.swift
@@ -9,7 +9,7 @@ let NavigationEventTypeRouteRetrieval = "mobile.performance_trace"
  */
 public protocol EventsManagerDataSource: AnyObject {
     var routeProgress: RouteProgress { get }
-    var router: Router! { get }
+    var router: Router { get }
     var desiredAccuracy: CLLocationAccuracy { get }
     var locationProvider: NavigationLocationManager.Type { get }
 }
@@ -39,6 +39,12 @@ open class NavigationEventsManager {
             _dataSource = newValue
         }
     }
+    
+    /**
+     Optional application metadata that that can help Mapbox more reliably diagnose problems that occur in the SDK.
+     For example, you can provide your application’s name and version, a unique identifier for the end user, and a session identifier.
+    */
+    public var userInfo: [String: String?]? = nil
     
     /**
      Indicates whether the application depends on MapboxNavigation in addition to MapboxCoreNavigation.
@@ -110,6 +116,7 @@ open class NavigationEventsManager {
         var event = NavigationEventDetails(dataSource: dataSource, session: sessionState, defaultInterface: usesDefaultUserInterface)
         event.event = MMEEventTypeNavigationCancel
         event.arrivalTimestamp = sessionState.arrivalTimestamp
+        event.appMetadata = userInfo
         
         let validRating: Bool = (rating >= MMEEventsManager.unrated && rating <= 100)
         assert(validRating, "MMEEventsManager: Invalid Rating. Values should be between \(MMEEventsManager.unrated) (none) and 100.")
@@ -131,6 +138,8 @@ open class NavigationEventsManager {
         var event = PerformanceEventDetails(event: NavigationEventTypeRouteRetrieval, session: sessionState, createdOn: sessionState.currentRoute.responseEndDate)
         event.counters.append(PerformanceEventDetails.Counter(name: "elapsed_time",
                                                               value: responseEndDate.timeIntervalSince(fetchStartDate)))
+        event.appMetadata = userInfo
+        
         if let routeIdentifier = sessionState.currentRoute.routeIdentifier {
             event.attributes.append(PerformanceEventDetails.Attribute(name: "route_uuid", value: routeIdentifier))
         }
@@ -142,6 +151,7 @@ open class NavigationEventsManager {
 
         var event = NavigationEventDetails(dataSource: dataSource, session: sessionState, defaultInterface: usesDefaultUserInterface)
         event.event = MMEEventTypeNavigationDepart
+        event.appMetadata = userInfo
         return event
     }
     
@@ -150,8 +160,9 @@ open class NavigationEventsManager {
 
         var event = NavigationEventDetails(dataSource: dataSource, session: sessionState, defaultInterface: usesDefaultUserInterface)
         event.event = MMEEventTypeNavigationArrive
-        
         event.arrivalTimestamp = dataSource.router.rawLocation?.timestamp ?? Date()
+        event.appMetadata = userInfo
+
         return event
     }
     
@@ -165,11 +176,12 @@ open class NavigationEventsManager {
     
     func navigationFeedbackEvent(type: FeedbackType, description: String?) -> NavigationEventDetails? {
         guard let dataSource = dataSource, let sessionState = sessionState else { return nil }
-
-        var event = NavigationEventDetails(dataSource: dataSource, session: sessionState, defaultInterface: usesDefaultUserInterface)
+        
+        var event = NavigationEventDetails(dataSource: dataSource, session: sessionState, defaultInterface: usesDefaultUserInterface, userInfo: userInfo)
         event.event = MMEEventTypeNavigationFeedback
         
         event.userId = UIDevice.current.identifierForVendor?.uuidString
+        event.appMetadata = userInfo
         event.feedbackType = type.description
         event.description = description
         
@@ -186,6 +198,7 @@ open class NavigationEventsManager {
         var event = NavigationEventDetails(dataSource: dataSource, session: sessionState, defaultInterface: usesDefaultUserInterface)
         event.event = eventType
         event.created = timestamp
+        event.appMetadata = userInfo
         
         if let lastRerouteDate = sessionState.lastRerouteDate {
             event.secondsSinceLastReroute = round(timestamp.timeIntervalSince(lastRerouteDate))
@@ -316,7 +329,8 @@ open class NavigationEventsManager {
      
      @param type A `FeedbackType` used to specify the type of feedback
      @param description A custom string used to describe the problem in detail.
-     @return Returns a UUID used to identify the feedback event
+     @param userInfo An optional dictionary used to include application metadata.
+     @return Returns a UUID used to identify the feedback event.
      
      If you provide a custom feedback UI that lets users elaborate on an issue, you should call this before you show the custom UI so the location and timestamp are more accurate.
      
@@ -325,6 +339,12 @@ open class NavigationEventsManager {
     public func recordFeedback(type: FeedbackType = .general, description: String? = nil) -> UUID? {
         return enqueueFeedbackEvent(type: type, description: description)
     }
+//    public func recordFeedback(type: FeedbackType = .general, description: String? = nil, userInfo: [String: String?]? = nil) -> UUID? {
+//        if (userInfo != nil) {
+//            self.userInfo = userInfo
+//        }
+//        return enqueueFeedbackEvent(type: type, description: description)
+//    }
     
     /**
      Update the feedback event with a specific feedback identifier. If you implement a custom feedback UI that lets a user elaborate on an issue, you can use this to update the metadata.