+++ Sources/MapboxNavigation/OrnamentsController.swift
@@ -14,7 +14,7 @@ extension NavigationMapView {
         weak var navigationViewData: NavigationViewData!
         weak var eventsManager: NavigationEventsManager!
         
-        fileprivate var navigationView: NavigationView! {
+        fileprivate var navigationView: NavigationView {
             return navigationViewData.navigationView
         }
         
@@ -36,7 +36,7 @@ extension NavigationMapView {
             }
             set {
                 if let newPosition = newValue {
-                    navigationView?.floatingButtonsPosition = newPosition
+                    navigationView.floatingButtonsPosition = newPosition
                 }
             }
         }
@@ -119,7 +119,7 @@ extension NavigationMapView {
         }
         
         @objc func feedback(_ sender: Any) {
-            guard let parent = navigationViewData.containerViewController else { return }
+            let parent = navigationViewData.containerViewController
             let feedbackViewController = FeedbackViewController(eventsManager: eventsManager)
             feedbackViewController.detailedFeedbackEnabled = detailedFeedbackEnabled
             parent.present(feedbackViewController, animated: true)
@@ -181,7 +181,8 @@ extension NavigationMapView {
         }
         
         private func labelCurrentRoadFeature(at location: CLLocation) {
-            guard let router = navigationViewData.router,
+            let router = navigationViewData.router
+            guard
                   let stepShape = router.routeProgress.currentLegProgress.currentStep.shape,
                   !stepShape.coordinates.isEmpty,
                   let mapView = navigationMapView.mapView else {